[#16] FEAT: 유저 탈퇴에 따른 음원, 리뷰 및 관련 엔티티 목록 전체 삭제

+++ auth-service/src/main/kotlin/com/auth_service/domain/interfaces/UserManageController.kt
@@ -63,7 +63,7 @@ class UserManageController(
         CookieUtils.deleteAllCookies(request, response)
         val record = mapOf(
             "topic" to "user-deletion-topic",
-            "message" to mapOf("userId" to sub)
+            "message" to mapOf("userId" to sub.toLong())
         )
         post(
             url = "http://localhost:9000/api/kafka",
